Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic[email]==2.9.2
aiosmtplib==3.0.1
email-validator==2.2.0
alembic==1.14.0